login_docker/requirements.txt
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary
python-dotenv==0.14.0
SQLAlchemy==1.3.19
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn==20.0.4
authomatic
Flask-JWT
Flask-Mail